# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateFieldRequest:
    properties:
      area:
        type: number
      coordinates:
        $ref: '#/definitions/models.Location'
      location:
        type: string
      name:
        type: string
    required:
    - location
    - name
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.Location:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  models.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a rice monitoring application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Rice Monitor API
  version: "1.0"
paths:
...
  /fields:
    get:
      description: Get a list of all fields for the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all fields
      tags:
      - fields
    post:
      consumes:
      - application/json
      description: Create a new field for the user
      parameters:
      - description: Field object that needs to be added
        in: body
        name: field
        required: true
        schema:
          $ref: '#/definitions/models.CreateFieldRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new field
      tags:
      - fields
  /fields/{id}:
    delete:
      description: Delete a field by its ID
      parameters:
      - description: Field ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a field
      tags:
      - fields
    get:
      description: Get a single field by its ID
      parameters:
      - description: Field ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a field by ID
      tags:
      - fields
    put:
      consumes:
      - application/json
      description: Update an existing field
      parameters:
      - description: Field ID
        in: path
        name: id
        required: true
        type: string
      - description: Field object that needs to be updated
        in: body
        name: field
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a field
      tags:
      - fields
...
schemes:
- http
- https
swagger: "2.0"